Refactor match controller into modular hooks
